Used Sheena's new storybook experimentation and refactored everything including logic. removed old files.
